refactor Page component fix paths
+ fix file paths to components
+ simplify some presenter components & HOC
