+++ src/frontend/web_application/src/layouts/Page/components/CallToAction/components/ComposeButton/presenter.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import { withTranslator } from '@gandi/react-translate';
 import ActionButton from '../ActionButton';
-import Icon from '../../../../../Icon';
+import Icon from '../../../../../../components/Icon';
 
 const Presenter = ({ translator: { __ }, ...props }) => (
   <ActionButton {...props}>
