+++ src/frontend/web_application/src/layouts/Page/components/Header/index.js
@@ -1,0 +1,4 @@
+import { withTranslator } from '@gandi/react-translate';
+import Presenter from './presenter';
+
+export default withTranslator()(Presenter);
