--- src/frontend/web_application/src/layouts/Page/components/MainView/style.scss
@@ -1,11 +1,0 @@
-@import '../../../../styles/common';
-
-.l-body {
-  &__content {
-    // XXX: need a better fix for call-to-action override on mobile
-    margin-bottom: 160px;
-    padding-top: 45px;
-
-    background-color: $co-color__fg__back;
-  }
-}
