+++ src/frontend/web_application/src/layouts/Page/components/OffCanvas/index.jsx
@@ -1,15 +1,32 @@
-import React, { Component } from 'react';
-import Presenter from './presenter';
+import React, { Component, PropTypes } from 'react';
 
 class OffCanvas extends Component {
+  static propTypes = {
+    leftChildren: PropTypes.node.isRequired,
+    children: PropTypes.node.isRequired,
+  };
+
   componentDidMount() {
     // eslint-disable-next-line no-new
     new Foundation.OffCanvas(jQuery('#left_off_canvas'), {});
   }
 
   render() {
+    const { leftChildren, children } = this.props;
+
     return (
-      <Presenter {...this.props} />
+      <div className="off-canvas-wrapper">
+        <div className="off-canvas-wrapper-inner" data-off-canvas-wrapper>
+          <div
+            className="off-canvas position-left"
+            id="left_off_canvas"
+            data-off-canvas
+          >{leftChildren}</div>
+          <div className="off-canvas-content" data-off-canvas-content>
+            {children}
+          </div>
+        </div>
+      </div>
     );
   }
 }
