+++ src/frontend/web_application/src/layouts/Page/presenter.jsx
@@ -1,12 +1,11 @@
 import React, { Component, PropTypes } from 'react';
-import './style.scss';
-import MainView from './components/MainView';
 import OffCanvas from './components/OffCanvas';
 import Header from './components/Header';
 import Navigation from './components/Navigation';
 import NavigationAlt from './components/NavigationAlt';
 import NotificationCenter from './components/NotificationCenter';
 import CallToAction from './components/CallToAction';
+import './style.scss';
 
 class PageContainer extends Component {
   static propTypes = {
@@ -23,12 +22,17 @@ class PageContainer extends Component {
 
     return (
       <OffCanvas leftChildren={<NavigationAlt />}>
-        <MainView
-          header={<Header />}
-          nav={<Navigation />}
-          callToAction={<CallToAction />}
-          notification={<NotificationCenter />}
-        >{ children }</MainView>
+        <div className="l-body">
+          <Header />
+          <Navigation />
+          <section role="main">
+            <div className="l-body__content">
+              {children}
+            </div>
+          </section>
+          <CallToAction />
+          <NotificationCenter />
+        </div>
       </OffCanvas>
     );
   }
