+++ src/frontend/web_application/src/layouts/Page/style.scss
@@ -6,3 +6,15 @@ $co-font-path: '../../fonts';
 
 // --- Vendor ---
 @import '../../styles/vendor/bootstrap_foundation-sites';
+
+@import '../../styles/common';
+
+.l-body {
+  &__content {
+    // XXX: need a better fix for call-to-action override on mobile
+    margin-bottom: 160px;
+    padding-top: 45px;
+
+    background-color: $co-color__fg__back;
+  }
+}
